backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# LangChain
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0

# Data integrations
notion-client>=2.2.0
gspread>=6.1.0
google-auth>=2.29.0

# Utilities
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
